[FIX] product: pricelist applied on product categories
This is related to revision 2ebbcba33a7034d737e25c25af0e74dc76db844d

The above revision introduces a regression: the ability
to configure pricelist items applied on product categories.

If you configured a pricelist with items applied
on products categories, then the first item with a product
category matched all products, even if the products were
not in the according category.

This is because the condition checking that the product
is within the pricelist item product category has been
completely removed, by the above commit.

opw-666059