[FIX] mail: prevent crash when interacting with inactive partner
when a user is disaninactivated, it makes sense to disaninactivate his/her
corresponding res_partner.  This had the unfortunate effect of changing
the result of the channel_slot call for the direct messages with this
user.  As a result, the discuss interface was not happy.  This commit
does what it needs to keep the discuss interface happy.  That's what we
want, right?