[FIX] web: allow view with arabic date saving
When editing a form with the arabic language, the date is formatted in
arabic. But the version of the Moment.js library we use for this had
inconsistencies and parsing the arabic date could not work.

Hence, saving would not work since date field values would be invalid.

This commit cherry-pick http://momentjs.com/ commits which solves that:

  from version 1.8.4: https://github.com/moment/moment/commit/c9023b5
  from version 1.9.0: https://github.com/moment/moment/commit/bd328d5

closes #10634
opw-660453