[FIX] mail: announcement bar URl displayed on print
On printing, the announcement bar URL was displayed
in the printing result.
e.g.
`https://services.openerp.com/openerp-enterprise/ab/register/...`
was displayed at the top of the print.

This is related to the revision 50a4ca31f8de8bb427cb1d0f182322a3d1e04486.

This is due to this simple typo, the according CSS rule was
never applied because of it.