[FIX] base: ir.qweb: save assets attchment with correct references
Force the values of `res_id` and `res_model` when creating assets
attachments. This avoid getting wrong default values from the context.

When using the mail composer with a automatically created report (like
when sending a quotation by email), if the report asset must be saved,
the context used is the one from the mail composer, which contains a
`default_res_id` entry.