[IMP] Purchase: Prevent PO creation on Vendor Bills
On the field 'Add Purchase Order' on vendor bills, don't allow the
user to create a new purchase order.

The button 'Set to Done' is not clear. This button will prevent the
creation of new invoices for this purchase order. Done is not a 'final'
state and the naming is not clear. Fp proposes to rename the state
'Locked' and the button 'Lock Bills'.