[FIX] event_sale: event tickets domain in sale order line
From saas-6, setting `0` as maximum available seats
means there is an unlimited amount of tickts that
can be sold, as stated in the help message of the field:
```
Define the number of available tickets. If you have too much registrations you will
not BE able to sell tickets anymore. Set 0 to ignore this rule set as unlimited.
```

This domain is the same domain used than in
the QWeb views of `website_event_sale`.

opw-669355