[FIX] point_of_sale: export all missing classes from js modules
In the interest of being flexible and customizeable all of the defined
classes in the POS should be exported so people can easily inherit them
and change their behavior. A lot of the time you can get to them even if
they're not exported if you get creative enough, but noone should have
to resort to such ugly methods.

This commit does this for all of the modules except for
point_of_sale.popups. It has been omitted because currently it exports a
single object (PopupWidget). Modifying this to export an object of
objects could potentially break other unofficial modules which we do not
want in a stable release.