[FIX] base: creation of delivery or invoice partner
The image that is set based on the type should not be colorized as it has no
transparency (image_colorize method would fail).
The _get_default_image is called explicitely to pass False to colorize attribute.

A proper refactoring of the method _get_default_image has been done at 66d4de7
but for 9.0 versions, makes this minimal hack without changing signature method.

Fixes #11027